+++ proc-macro-definitions/src/part/html_definition.rs
@@ -293,7 +293,7 @@ impl<C: Configuration> HtmlDefinition<C> {
 		}
 		let children = quote_spanned! {child_stream.span()=>
 			&*#bump.alloc_try_with(
-				|| -> ::std::result::Result<_, ::#asteracea::error::GUIError> {
+				|| -> ::std::result::Result<_, ::#asteracea::error::Escalation> {
 					::std::result::Result::Ok([#child_stream])
 				}
 			)?