+++ proc-macro-definitions/src/trace_instrumentation.rs
@@ -2,15 +2,14 @@ use proc_macro2::{Span, TokenStream};
 use quote::{quote_spanned, ToTokens};
 use syn::{parse::Parse, spanned::Spanned, token::Brace, Token};
 use syn_mid::{Block, ItemFn};
-
 use crate::asteracea_ident;
 
-pub struct GuiTraced {
+pub struct Tracing {
 	function: ItemFn,
 	pub prefix: Option<TokenStream>,
 }
 
-impl Parse for GuiTraced {
+impl Parse for Tracing {
 	fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
 		Ok(Self {
 			function: input.parse()?,
@@ -19,7 +18,7 @@ impl Parse for GuiTraced {
 	}
 }
 
-impl ToTokens for GuiTraced {
+impl ToTokens for Tracing {
 	fn to_tokens(&self, tokens: &mut proc_macro2::TokenStream) {
 		ItemFn {
 			block: {
@@ -33,8 +32,13 @@ impl ToTokens for GuiTraced {
 				}).flatten();
 				let name = &self.function.sig.ident;
 				let inner = quote_spanned! {self.function.sig.output.span().resolved_at(Span::mixed_site())=>
-					let result: ::std::result::Result<_, ::#asteracea::error::GUIError> = (|| { #statements })();
-					result.map_err(|gui_error| gui_error.__Asteracea__with_traced_frame(::std::borrow::Cow::Borrowed(::std::stringify!(#prefix#joiner#name))))
+					let result: ::std::result::Result<_, ::#asteracea::error::Escalation> = (|| { #statements })();
+					result.map_err(|gui_error| gui_error.__Asteracea__with_traced_frame(::std::borrow::Cow::Borrowed(::std::concat!(
+						// Not stringifying in bulk avoids spaces between the tokens.
+						::std::stringify!(#prefix),
+						::std::stringify!(#joiner),
+						::std::stringify!(#name),
+					))))
 				};
 				Box::new(Block {
 					brace_token: Brace(Span::mixed_site()),